Fix: App.js을 route.js로 변경
