+++ src/component/Header.js
@@ -37,8 +37,6 @@ export const Header = (userName) => /* HTML */ {
 		</li>
 	  </ul>
 	</nav>`;
-  const logoutButton = document.getElementById("logout");
-  console.log(logoutButton);
   return /*html*/ `
 	<header class="bg-blue-600 text-white p-4 sticky top-0">
 	  <h1 class="text-2xl font-bold">항해플러스</h1>
