+++ src/lib/render.js
@@ -1,9 +1,7 @@
-import { App } from "../App";
 import { route } from "./route";
 
 const root = document.getElementById("root");
 
 export const render = () => {
-  root.innerHTML = App();
-  route();
+  root.innerHTML = route();
 };
