+++ src/lib/route.js
@@ -1,14 +1,8 @@
-import { render } from "./render";
-
-const root = document.getElementById("root");
+import { MainPage, ProfilePage, LoginPage, ErrorPage } from "../page";
 
 export const route = () => {
-  root.querySelectorAll("a").forEach((el) => {
-    el.addEventListener("click", (e) => {
-      e.preventDefault();
-      const nextPathName = e.target.href.replace(location.origin, "");
-      history.pushState(null, "", nextPathName);
-      render();
-    });
-  });
+  if (location.pathname === "/") return MainPage();
+  if (location.pathname === "/profile") return ProfilePage();
+  if (location.pathname === "/login") return LoginPage();
+  return ErrorPage();
 };
