+++ src/main.js
@@ -14,14 +14,11 @@ root.addEventListener("submit", function (e) {
 
   if (e.target && e.target.id === "login-form") {
     let username = document.getElementById("username")?.value;
-    setUser({ username, bio: "", email: "" });
+    setUser({ username });
     history.pushState(null, "", "profile");
     render();
   }
-});
 
-root.addEventListener("submit", function (e) {
-  e.preventDefault();
   if (e.target && e.target.id === "profile-form") {
     let username = document.getElementById("username")?.value;
     let bio = document.getElementById("bio")?.value;
@@ -31,6 +28,13 @@ root.addEventListener("submit", function (e) {
 });
 
 root.addEventListener("click", function (e) {
+  if (e.target && e.target.nodeName == "A") {
+    e.preventDefault();
+    const nextPathName = e.target.href.replace(location.origin, "");
+    history.pushState(null, "", nextPathName);
+    render();
+  }
+
   if (e.target && e.target.id === "logout") {
     clearUser();
     history.pushState(null, "", "login");
